requirements.txt
streamlit>=1.30.0
pymupdf>=1.23.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
